moved styles out and add a constant file
