+++ src/components/ColorPickerForm/index.js
@@ -3,23 +3,7 @@ import { withStyles } from "@material-ui/styles";
 import { Button } from "@mui/material";
 import { ChromePicker } from "react-color";
 import { ValidatorForm, TextValidator } from "react-material-ui-form-validator";
-
-const styles = {
-  picker: {
-    width: "100% !important",
-    marginTop: "2rem",
-  },
-  addColor: {
-    width: "100%",
-    padding: "1rem !important",
-    marginTop: "1rem !important",
-    fontSize: "1rem !important",
-  },
-  colorNameInput: {
-    width: "100%",
-    height: "70px",
-  },
-};
+import styles from "../../styles/ColorPickerFormStyles";
 
 class ColorPickerForm extends Component {
   constructor(props) {
