+++ src/components/DraggableColorBox/index.js
@@ -2,39 +2,7 @@ import React from "react";
 import { SortableElement } from "react-sortable-hoc";
 import { withStyles } from "@material-ui/styles";
 import DeleteIcon from "@mui/icons-material/Delete";
-
-const styles = {
-  root: {
-    width: "20%",
-    height: "25%",
-    margin: "0 auto",
-    display: "inline-block",
-    position: "relative",
-    cursor: "pointer",
-    marginBottom: "-3.5px",
-    marginTop: "20px",
-    "&:hover svg": {
-      color: "white",
-      transform: "scale(1.5)",
-    },
-  },
-  boxContent: {
-    position: "absolute",
-    width: "100%",
-    left: "0px",
-    bottom: "0px",
-    padding: "10px",
-    color: "rgba(0, 0, 0, 0.5)",
-    letterSpacing: "1px",
-    textTransform: "uppercase",
-    fontSize: "12px",
-    display: "flex",
-    justifyContent: "space-between",
-  },
-  deleteIcon: {
-    transition: "all 0.3s ease-in-out !important",
-  },
-};
+import styles from "../../styles/DraggableColorBoxStyles";
 
 const DraggableColorBox = SortableElement((props) => {
   const { classes, handleClick, name, color } = props;
