+++ src/styles/NewPaletteFormStyles.js
@@ -1,0 +1,57 @@
+import { styled } from "@mui/material/styles";
+import { DRAWER_WIDTH } from "../utils/appConstants";
+
+export const Main = styled("main", {
+  shouldForwardProp: (prop) => prop !== "open",
+})(({ theme, open }) => ({
+  flexGrow: 1,
+  height: "calc(100vh - 64px)",
+  padding: theme.spacing(3),
+  transition: theme.transitions.create("margin", {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  marginLeft: `-${DRAWER_WIDTH}px`,
+  ...(open && {
+    transition: theme.transitions.create("margin", {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen,
+    }),
+    marginLeft: 0,
+  }),
+}));
+
+export const DrawerHeader = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  padding: theme.spacing(0, 1),
+  ...theme.mixins.toolbar,
+  justifyContent: "flex-end",
+}));
+
+export const styles = {
+  container: {
+    width: "90%",
+    height: "100%",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  drawerPaper: {
+    width: `${DRAWER_WIDTH}px`,
+    flexShrink: 0,
+    "& .MuiDrawer-paper": {
+      width: DRAWER_WIDTH,
+      boxSizing: "border-box",
+      display: "flex",
+      alignItems: "center",
+    },
+  },
+  buttons: {
+    width: "100%",
+  },
+  button: {
+    width: "50%",
+  },
+};
